# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poll Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SummaryDateRange = {
  from: string;
  to: string;
};

export interface Summary {
  total_sessions: number;
  total_responses: number;
  avg_delivery: number;
  avg_content: number;
  avg_combined: number;
  subjects: string[];
  programs: string[];
  date_range: SummaryDateRange;
  total_instructors: number;
  total_topics: number;
}

export interface SubjectSession {
  upload_id: number;
  session_date: string;
  week_number: string;
  total_responses: number;
  avg_delivery: number;
  avg_content: number;
  instructor?: string | null;
  topic?: string | null;
}

export interface TrendPoint {
  session_date: string;
  week: string;
  avg_delivery: number;
  avg_content: number;
  responses: number;
  subject?: string;
}

export interface SubjectTrendPoint {
  week: string;
  delivery: number;
  content: number;
}

export interface SubjectAnalytics {
  subject: string;
  total_sessions: number;
  total_responses: number;
  avg_delivery: number;
  avg_content: number;
  nps: number;
  trend: SubjectTrendPoint[];
  instructors?: string[];
}

export interface CohortSubjectRating {
  [key: string]: {
    avg_delivery: number;
    avg_content: number;
  };
}

export interface CohortAnalytics {
  cohort: string;
  total_responses: number;
  avg_delivery: number;
  avg_content: number;
  by_subject: CohortSubjectRating;
}

export interface Distribution {
  "1": number;
  "2": number;
  "3": number;
  "4": number;
  "5": number;
  total: number;
  nps: number;
}

export interface DistributionBucket {
  "1": number;
  "2": number;
  "3": number;
  "4": number;
  "5": number;
}

export interface OverallDistribution {
  delivery: DistributionBucket;
  content: DistributionBucket;
  total_responses: number;
  nps_delivery: number;
  nps_content: number;
}

export interface ThemeCount {
  theme: string;
  count: number;
}

export interface SentimentBreakdown {
  positive: number;
  negative: number;
  suggestion: number;
  neutral: number;
}

export interface SentimentPercent {
  positive: number;
  negative: number;
  suggestion: number;
  neutral: number;
}

export interface FeedbackAnalysis {
  total_feedback: number;
  useful_feedback: number;
  sentiment: SentimentBreakdown;
  themes: ThemeCount[];
  top_negative: string[];
  top_suggestions: string[];
  top_positive: string[];
}

export interface FeedbackOverview {
  total_useful: number;
  sentiment_counts: SentimentBreakdown;
  sentiment_percent: SentimentPercent;
  top_themes: ThemeCount[];
}

export interface UploadHistory {
  id: number;
  subject: string;
  session_date: string;
  program_name: string;
  week_number: string;
  total_responses: number;
  avg_combined_rating: number;
  upload_timestamp: string;
  session_type: string;
  meeting_topic: string;
  format_version?: string;
  instructor?: string | null;
  topic?: string | null;
}

export interface UploadResult {
  success: boolean;
  upload_id: number;
  subject: string;
  total_responses: number;
  avg_delivery_rating: number | null;
  avg_content_rating: number | null;
  message?: string;
  format?: string;
  instructor?: string | null;
  topic?: string | null;
}

export interface UploadError {
  success: boolean;
  error: string;
  message: string;
  original_upload_date?: string;
}

export interface InstructorSummary {
  instructor: string;
  total_sessions: number;
  subjects: string[];
  total_responses: number;
  avg_rating: number;
  avg_delivery: number;
  avg_content: number;
}

export interface InstructorByWeek {
  session_date: string;
  subject: string;
  avg_rating: number;
  responses: number;
}

export type InstructorDetailBySubject = {
  [key: string]: {
    avg?: number;
    sessions?: number;
  };
};

export type InstructorDetailFeedbackSummary = {
  positive?: number;
  negative?: number;
  suggestion?: number;
  top_themes?: string[];
};

export interface InstructorDetail {
  instructor: string;
  total_sessions: number;
  total_responses: number;
  avg_rating: number;
  avg_delivery?: number;
  avg_content?: number;
  by_subject: InstructorDetailBySubject;
  by_week: InstructorByWeek[];
  feedback_summary: InstructorDetailFeedbackSummary;
}

export interface TopicSummary {
  topic: string;
  subject: string;
  instructor?: string | null;
  sessions: number;
  total_responses: number;
  avg_rating: number;
  feedback_count: number;
}

export type TopicDetailDistribution = { [key: string]: number };

export type TopicDetailFeedback = {
  useful?: number;
  sentiment?: SentimentBreakdown;
  top_issues?: string[];
  suggestions?: string[];
};

export interface TopicDetail {
  topic: string;
  subject: string;
  instructor?: string | null;
  session_date: string;
  total_responses: number;
  avg_delivery: number;
  avg_content: number;
  distribution: TopicDetailDistribution;
  nps: number;
  feedback: TopicDetailFeedback;
}

export type GetSubjectSessionsParams = {
  subject: string;
};

export type GetTrendsParams = {
  subject?: string;
};

export type GetDistributionParams = {
  subject?: string;
  type?: GetDistributionType;
  upload_id?: number;
};

export type GetDistributionType =
  (typeof GetDistributionType)[keyof typeof GetDistributionType];

export const GetDistributionType = {
  delivery: "delivery",
  content: "content",
} as const;

export type GetFeedbackParams = {
  subject?: string;
  upload_id?: number;
};

export type UploadCsvBody = {
  file: Blob;
  program_name: string;
  session_type?: string;
  cohort?: string;
  semester: string;
  week_number: string;
  instructor?: string;
  topic?: string;
};

export type PatchUploadBody = {
  instructor?: string | null;
  topic?: string | null;
};

export type PatchUpload200 = {
  success: boolean;
};

export type GetTopicsParams = {
  subject?: string;
  instructor?: string;
};