# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Poll Analytics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload:
    post:
      operationId: uploadCsv
      tags: [upload]
      summary: Upload a Zoom poll CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                program_name:
                  type: string
                session_type:
                  type: string
                cohort:
                  type: string
                semester:
                  type: string
                week_number:
                  type: string
                instructor:
                  type: string
                topic:
                  type: string
              required:
                - file
                - program_name
                - semester
                - week_number
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Upload error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadError"
  /upload/{id}:
    patch:
      operationId: patchUpload
      tags: [upload]
      summary: Update instructor and topic for an upload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instructor:
                  type: string
                  nullable: true
                topic:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
components:
  schemas:
...
    UploadResult:
      type: object
      properties:
        success:
          type: boolean
        upload_id:
          type: integer
        subject:
          type: string
        total_responses:
          type: integer
        avg_delivery_rating:
          type: number
          nullable: true
        avg_content_rating:
          type: number
          nullable: true
        message:
          type: string
        format:
          type: string
        instructor:
          type: string
          nullable: true
        topic:
          type: string
          nullable: true
      required:
        - success
        - upload_id
        - subject
        - total_responses
        - avg_delivery_rating
        - avg_content_rating
    UploadError:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        message:
          type: string
        original_upload_date:
          type: string
      required:
        - success
        - error
        - message
...