# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poll Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  success: boolean;
  upload_id: number;
  subject: string;
  total_responses: number;
  avg_delivery_rating: number | null;
  avg_content_rating: number | null;
  message?: string;
  format?: string;
  instructor?: string | null;
  topic?: string | null;
}